// src/models/public/GoalReportValues.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { GoalReportsGoalReportId } from './GoalReports';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.goal_report_values */
export type GoalReportValuesGoalReportValueId = string & { __brand: 'public.goal_report_values' };

/**
 * Represents the table public.goal_report_values
 * value-ing the success of goals
 */
export default interface GoalReportValues {
  goal_report_value_id: GoalReportValuesGoalReportValueId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  goal_report_id: GoalReportsGoalReportId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer values */
  value_integer: number | null;

  /** Used for numeric values */
  value_numeric: number | null;

  /** Used for text values */
  value_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.goal_report_values
 * value-ing the success of goals
 */
export interface GoalReportValuesInitializer {
  /** Default value: uuid_generate_v7() */
  goal_report_value_id?: GoalReportValuesGoalReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  goal_report_id?: GoalReportsGoalReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.goal_report_values
 * value-ing the success of goals
 */
export interface GoalReportValuesMutator {
  goal_report_value_id?: GoalReportValuesGoalReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  goal_report_id?: GoalReportsGoalReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}